// web/src/api/generated/models/MilestoneUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MilestoneUpdate
 */
export interface MilestoneUpdate {
    /**
     * 
     * @type {string}
     * @memberof MilestoneUpdate
     */
    body?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneUpdate
     */
    happenedOn?: Date;
    /**
     * 
     * @type {string}
     * @memberof MilestoneUpdate
     */
    title?: string;
}

/**
 * Check if a given object implements the MilestoneUpdate interface.
 */
export function instanceOfMilestoneUpdate(value: object): value is MilestoneUpdate {
    return true;
}

export function MilestoneUpdateFromJSON(json: any): MilestoneUpdate {
    return MilestoneUpdateFromJSONTyped(json, false);
}

export function MilestoneUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): MilestoneUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'body': json['body'] === undefined ? undefined : json['body'] === null ? null : json['body'],
        'happenedOn': json['happenedOn'] == null ? undefined : (new Date(json['happenedOn'])),
        'title': json['title'] == null ? undefined : json['title'],
    };
}

export function MilestoneUpdateToJSON(json: any): MilestoneUpdate {
    return MilestoneUpdateToJSONTyped(json, false);
}

export function MilestoneUpdateToJSONTyped(value?: MilestoneUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'body': value['body'],
        'happenedOn': value['happenedOn'] == null ? value['happenedOn'] : value['happenedOn'].toISOString().substring(0,10),
        'title': value['title'],
    };
}